requirements.txt
PyAutoGUI==0.9.54
pyscreenshot==3.1
pytesseract==0.3.10
pynput==1.7.7
opencv-python==4.10.0.84
numpy==1.26.4
google.generativeai
python-dotenv